samjobBackend/requirements.txt
about-time==4.2.1
absl-py==1.4.0
alive-progress==3.1.4
argcomplete==2.0.0
asgiref==3.7.2
astroid==2.15.4
asttokens==2.1.0
astunparse==1.6.3
backcall==0.2.0
Beaker==1.12.1
beautifulsoup4==4.12.2
blivet==3.7.1
blivet-gui==2.4.1
Brlapi==0.8.4
bs4==0.0.1
cachetools==5.3.1
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
conda==23.3.1
conda-package-handling==2.0.2
conda_package_streaming==0.7.0
cryptography==37.0.2
cupshelpers==1.0
cytoolz==0.12.0
dasbus==1.7
dbus-python==1.3.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
distro==1.8.0
Django==4.2.2
django-cors-headers==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
docutils==0.19
executing==1.2.0
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.8.2
flatbuffers==23.5.26
fros==1.1
frozendict==2.3.5
gast==0.4.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gpg==1.17.1
grapheme==0.6.0
grpcio==1.54.2
h5py==3.8.0
humanize==3.13.1
idna==3.4
iotop==0.6
ipython==8.14.0
isort==5.12.0
jax==0.4.12
jedi==0.18.2
joblib==1.2.0
keras==2.12.0
langtable==0.0.62
lazy-object-proxy==1.9.0
libclang==16.0.0
libcomps==0.1.18
lxml==4.9.2
Mako==1.2.3
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mccabe==0.7.0
ml-dtypes==0.2.0
mypy==1.4.0
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
nftables==0.1
numpy==1.23.5
oauthlib==3.2.2
olefile==0.46
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.3
parso==0.8.3
Paste==3.5.2
pexpect==4.8.0
pickleshare==0.7.5
pid==2.2.3
Pillow==9.4.0
platformdirs==2.6.0
pluggy==1.0.0
ply==3.11
productmd==1.36
progress==1.6
progressbar2==4.2.0
prompt-toolkit==3.0.38
protobuf==4.23.2
ptyprocess==0.7.0
pure-eval==0.2.2
pwquality==1.4.5
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.23.0
pycosat==0.6.4
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.45.2
pyenchant==3.2.2
pyFirmata==1.1.0
pyFirmata2==2.4.2
Pygments==2.15.1
PyGObject==3.44.1
PyJWT==2.8.0
pykickstart==3.47
pylint==2.17.4
pyOpenSSL==21.0.0
pyparsing==3.0.9
pyparted==3.12.0
pyserial==3.5
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-meh==0.51
python-utils==3.6.0
python3-openid==3.2.0
pytz==2023.3
pyudev==0.24.0
pyxdg==0.27
PyYAML==6.0
regex==2023.6.3
requests==2.28.2
requests-file==1.5.1
requests-ftp==0.3.1
requests-oauthlib==1.3.1
rpm==4.18.1
rsa==4.9
ruamel.yaml==0.17.24
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scipy==1.10.1
selinux==3.5
Send2Trash==1.8.0
sepolicy==3.5
setools==4.4.2
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sos==4.5.1
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
systemd-python==235
Tempita==0.5.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
thonny==4.0.2
threadpoolctl==3.1.0
tomlkit==0.11.4
toolz==0.12.0
tqdm==4.65.0
traitlets==5.9.0
typed-ast==1.5.5
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.21.1
wcwidth==0.2.6
Werkzeug==2.3.6
wrapt==1.14.1
zstandard==0.21.0
